lib/generators/app/templates/server/requirements.txt
Django==1.9.4
dj-database-url==0.4.1
django-flatblocks==0.9.2
funcy==1.7.1
gunicorn==19.4.5
ipdb==0.9.3
ipython==4.2.0
psycopg2==2.6.1
wheel==0.26.0
whitenoise==3.0

django-jfu==2.0.9
django-geoposition==0.2.2
django-grappelli==2.8.1
django-webtest==1.7.9
factory-boy==2.4.1
fake-factory==0.5.7
cerberus==0.9.2
django-wysiwyg-redactor==0.4.9
django-admin-sortable2==0.6.3
git+https://github.com/Brogency/django-db-mailer.git
django-picklefield==0.3.2
django-constance==1.1.2
djangorestframework==3.3.3
rest-framework-generic-relations==1.1.0
django-filter==0.13.0
xlrd==0.9.4
django-debug-toolbar==1.5
django-rest-swagger==0.3.6
django-cors-headers==0.12
mock==2.0.0
raven==5.26.0
django-settings-export==1.1.0
sorl-thumbnail==12.4a1
transliterate==1.7.6

celery==3.1.23
redis==2.10.5
django-redis==4.4.4

requests==2.9.1
lxml==3.6.0
Pillow==3.2.0
coverage==4.0.3
